+++ ontask/agent/utils/authenticate.py
@@ -3,13 +3,11 @@ import os
 import requests
 import logging
 import environ
-from colorama import Fore
-import pandas as pd
 
 from ontask.agent.canvas_api.c3l import C3L
 # Setting up logging
 logging.basicConfig(level=logging.INFO)
-
+env = environ.Env()
 env = environ.Env()
 
 class agent:
@@ -31,13 +29,17 @@ class agent:
     def get_quizes(self):
         pass
 
+config = {
+    'canvas_base_url': env("CANVAS_BASE_URL"),
+    'canvas_api_token': env("CANVAS_API_TOKEN"),
+    'ontask_base_url': env("ONTASK_BASE_URL"),
+    'ontask_api_token': env("ONTASK_API_TOKEN")
+}
 
-# Load config.json
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-CONFIG_PATH = os.path.join(BASE_DIR, 'config.json')
-
-with open(CONFIG_PATH, 'r') as file:
-    config = json.load(file)
+print("CANVAS_BASE_URL", env("CANVAS_BASE_URL", default=""))
+print("CANVAS_API_TOKEN", env("CANVAS_API_TOKEN", default=""))
+print("ONTASK_BASE_URL", env("ONTASK_BASE_URL", default=""))
+print("ONTASK_API_TOKEN", env("ONTASK_API_TOKEN", default=""))
 
 def authenticate_canvas():
     headers = {