+++ ChangeLog.md
@@ -1,3 +1,24 @@
+<<<<<<< HEAD
+=======
+# 10.6 (2023-12-5)
+
+## Added
+
+- New functionality to schedule Canvas Email action execution repeatedly
+
+>>>>>>> origin/master
+# 10.5 (2023-11-25)
+
+## Fixed
+
+- Error in merge operation when new dataframe does not have extra columns but only extra rows.
+
+## Changed
+
+- Upgraded libraries
+
+- Adjusted usage of Selenium
+
 # 10.4 (2023-08-16)
 
 ## Changed